feat(pgx) + interface: Repository
While it would be possible to import and use the *Queries type directly
in our GraphQL resolvers, taking this approach would make testing the
resolver code difficult later on. The resolvers would end up coupled
with the database code and we would most likely have no choice but to
rely on integration tests in order to test them. We will opt for a
different approach and create a custom Repository interface to serve as,
an interface to the sqlc-generated code. The rest of our application
will only communicate with the database through that interface, which
means we will be able to use the mock version of the Repository in our
tests.

The Repository interface is a subset of query methods generated for us
by sqlc along with a couple of custom methods. It only exposes a subset
of the queries because some of the queries we have created were never
meant to be used as stand-alone operations so the rest of our
application neither has to nor should be able to access them directly.

You'll notice that two methods of the Repository interface - CreateBook
and UpdateBook - have signatures that differ from those of their
namesakes from the queries.sql.go file. The generated CreateBook and
UpdateBook methods are only concerned with making changes to the books
table, which means they don't know anything about the authors of a given
book. The complete process of creating or updating a book must include
creating or updating the associations between the book and its authors
and as such it will have to take the form of a sequence of database
operations wrapped in a transaction. We will need to create custom
implementations for these methods.