feat(pgx) + repoSvc.WithTx()
Embedding *Queries in the repoSvc struct means that it automatically
implements those methods of the Repository interface that were derived
from the *Queries struct in the first place. This means we only need to
create custom implementations of the CreateBook and UpdateBook methods.
Since both of these methods will involve database transactions let's
start by creating a helper method for managing them.