build(gqlg) + config: wrap nullable inputs in type graphql.Omittable
This allows us to check if an input to a resolver has been explicitly
set to null or omitted from the request. Without this, the client would
need to specify every field of an object during an update mutation, else
the unspecified nullable fields would be set to null, erasing exsiting
data in the database.