*** Implementing Data Loaders ***
At this point, we have a working database and graphql API to interact with the
database. However, the current implementation suffers from the "N+1" problem.

The N+1 problem in GraphQL occurs when a query retrieves a list of items and
then makes additional queries to fetch related data for each item, leading to a
large number of unnecessary database queries.

For example, a query that fetches a list of authors might result in one query to
fetch the authors and then an additional query for each author to fetch their
agents, leading to a total of N+1 queries if there are N authors.

Since we have postgres configured to log queries to stderr within the compose
file, you can see this in action by creating an agent and then several authors.
If you then execute the "authors" query, you will see one query for the list of
authors, and a subsequent query for each created author to obtain the agent ID.

The solution to this problem is to implement data loaders. A dataloader is
defined as a request-scoped method of caching and batching data-fetching
queries. Utilizing a dataloader, the above-described query would result in the
following sequence of events:

a. A dataloader instance would be created at the beginning of the request.

b. Instead of calling the database directly, the individual invocations of the
authorResolver.Agent method (one for each author) would instead call the Load
method of the dataloader instance with the id of the required agent.

c. The dataloader instance, meanwhile, would first wait for a few milliseconds
to collect the agent ids for its database query, and then would make the query
and fetch all required agents at once.

d. Having retrieved the data, the dataloader instance would distribute it back
to the respective resolver methods.

e. Finally, upon the completion of the request, the dataloader instance would
fall out of scope to be garbage-collected.

Dataloaders are request-scoped, which means they only exist for the duration of
a request. Given this property, it makes sense to store them in the request
context. Dataloaders, furthermore, are specific to the types of values they
process, which means we will need to implement different dataloaders for
different resolvers.

In order to be able to use dataloaders in our application we first have to
figure out how to:

1. Initialize the dataloaders and store them as a context value for each
incoming request.

2. Make the request-scoped dataloader instances available to the resolvers.