feat(dataloaders) + type: loaders with constructor
The exported Loaders type is an empty struct to begin with, but it will
soon hold the various dataloader implementations used by our
application.

The newLoaders function will be used to instantiate the dataloaders. The
function takes the context of the current request (ctx) and the
repository (pgx.Repository) as arguments, the latter because the
dataloaders will need to query the database.