build(dataloaden) + generate: AgentLoader
For the authorResolver.Agent method we will need a dataloader which
takes int64 input values and returns values of *pgx.Agent type. In other
words, for every agent_id value the dataloader will need to return a
pointer to an Agent struct. We add the command required to generate a
dataloader for this combination of types as a comment to the
dataloaders/dataloaders.go file.